Updating classes and project structure to be in the standard of ports and adapters architecure by FIAP (#31)
* Updating classes and project structure to be in the standard of ports and adapters scructure since FIAP sent a feedback about this

* Removing getters

* Removing getters and setter from Product

* Removing getters and setter from order

* Removing getters and setter from Customer

* Removing getters and setter from Cart Item

* Removing getters and setter from Cart

+++ src/main/java/br/com/fiap/controlepedidos/adapters/driver/apirest/controllers/CheckoutController.java
@@ -1,5 +1,8 @@
 package br.com.fiap.controlepedidos.adapters.driver.apirest.controllers;
 
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RestController;
+
 import br.com.fiap.controlepedidos.adapters.driver.apirest.contract.CheckoutApi;
 import br.com.fiap.controlepedidos.adapters.driver.apirest.dto.in.DoCheckoutRequestDTO;
 import br.com.fiap.controlepedidos.adapters.driver.apirest.dto.in.MercadoPagoQRCodePaymentCallbackDTO;
@@ -7,8 +10,6 @@ import br.com.fiap.controlepedidos.adapters.driver.apirest.dto.out.MercadoPagoQr
 import br.com.fiap.controlepedidos.adapters.driver.apirest.dto.out.PaymentDataResponseDTO;
 import br.com.fiap.controlepedidos.core.application.services.checkout.ConfirmQRCodePaymentOrder;
 import br.com.fiap.controlepedidos.core.application.services.checkout.StartCheckoutService;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 public class CheckoutController implements CheckoutApi {
@@ -28,6 +29,6 @@ public class CheckoutController implements CheckoutApi {
 
     @Override
     public ResponseEntity<MercadoPagoQrCodePaymentCallbackResponseDTO> confirmPayment(MercadoPagoQRCodePaymentCallbackDTO callbackDetails) {
-        return ResponseEntity.ok(MercadoPagoQrCodePaymentCallbackResponseDTO.fromDomain(confirmQRCodePaymentOrder.confirmQrCodePayment(callbackDetails.orderId(), callbackDetails.paidValue().intValue() )));
+        return ResponseEntity.ok(MercadoPagoQrCodePaymentCallbackResponseDTO.fromDomain(confirmQRCodePaymentOrder.confirmQrCodePayment(callbackDetails.orderId(), callbackDetails.paidValue() )));
     }
 }
